+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This changelog's format is based on [keep a changelog 1.0.0](https://keepachange
 ### Added
 - population-counter gate
 - ladder constraint encoder
+- ClauseConsumer decorator for adding activation literals
 - performance tests for the bitvector multiplication gate encoder and the ltseq cardinality
   constraint encoder